Conflicts resolved

+++ src/main/java/com/rabzelilov/spring_test/dal/repository/DnaRepository.java
@@ -3,5 +3,5 @@ package com.rabzelilov.spring_test.dal.repository;
 import com.rabzelilov.spring_test.dal.entity.Dna;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DnaRepository extends JpaRepository<Dna,Integer> {
+public interface DnaRepository extends JpaRepository<Dna,Long> {
 }
